pyproject.toml
[project]
name = "add-gym"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.15"
dependencies = [
    "genesis-world==0.3.10",
    "hydra-core>=1.3.2",
    "ipython>=8.37.0",
    "s3fs>=2026.1.0",
    "tensorboard>=2.20.0",
    "torch>=2.8.0",
]

[project.optional-dependencies]
mjwarp = [
    "mujoco-warp>=0.0.2",
]
